Make InteractiveCrawler abstract and require that subclasses provide a UI panel.
It makes much more sense to ask the interactive crawlers to provide the correct
UI for themselves, so this patch does that.

As a result certain other UI classes need to change slightly. ControlPanel now
asks its crawler for the panel if it's interactive. As a result it no longer
needs to know anything about the size of the domain so its constructor is
modified to reflect this.